refactor: move hide/show methods to StatycView; remove jQuery signs; DynamicView will not return an array of elements
DynamicView will return an array instead of the first child, because the
template can contain more that one element, which should be inserted
into the new DOM container.

Also the temporary container, where DynamicView will be rendered, can be
changed. For example to properly render TD elements the temporary
container should be TR. It will not correctly rendered in DIV element.