update: set view as a context of a template
Before template accepted only the model as a context for processing.
The idea behind that was, that template is a representation of a model,
nothing more.
But there can be a situation when only for some of representation(s)
small logic is required to process the model data and give a single
view of it. That is why the view object is passed to the template
engine as a context. So now inside the template the methods of view can
be used for some data processing.
With this the relation between template and view is more direct.